fix: return correct xml type for bool
As discussed nicely in the issue
https://github.com/vinitkumar/json2xml/issues/118, the correct bool
value is never returned as the bool is also a subtype of numbers.Number
and int.

Here, hence, the check for bool is done first because while the above
condition might be true for bool, no other type can say the same for
bool check. i.e. Numbers won't be an instance of bool and so would be
int.

A test is added to guard against the future regression on this as well.

- Github Issue: Resolves #118

Authored-by: Vinit Kumar <vinit.kumar@kidskonnect.nl>
Signed-off-by: Vinit Kumar <vinit.kumar@kidskonnect.nl>