fix: globally shared ids list that gets mutated
We have a ids variable that stores a list of unique ids. The issue is
that it is globally shared and can mutate across tests. This caused a
failure in one of the test that shouldn't contain id but it does in this
case.

When the global ids is removed and put into the function where it is
used and the reference of ids in other functions properly instantiated,
we can see that the crash is gone.

- Github Issue: #150

Authored-by: Vinit Kumar <vinit.kumar@kidskonnect.nl>
Signed-off-by: Vinit Kumar <vinit.kumar@kidskonnect.nl>